Watch App finishing touches. Added integration with server. Added a slightly better number input.

+++ MeterBeater WatchKit 1 Extension/Request.swift
@@ -10,12 +10,12 @@ import Foundation
 import CoreLocation
 
 class Request {
-    var timeInMinutes: Float
+    var timeInMinutes:Int
     var latitude: CLLocationDegrees
     var longitude: CLLocationDegrees
     var coordinate: CLLocationCoordinate2D
     
-    init(timeInMinutes: Float, latitude: CLLocationDegrees, longitude: CLLocationDegrees, coordinate: CLLocationCoordinate2D) {
+    init(timeInMinutes: Int, latitude: CLLocationDegrees, longitude: CLLocationDegrees, coordinate: CLLocationCoordinate2D) {
         self.timeInMinutes = timeInMinutes
         self.latitude = latitude
         self.longitude = longitude
